multi-class perceptron language detection
Implemented training for multiple perceptrons to detect languages. Added epoch-based training with accuracy feedback.

Epochs by defualt are 7, but generally the project accepts epochs as first arg.

To admit, the test/train data is very random and not very well chosen, which is why usually unless a lot of sentences are input, the detection might be very inaccurate.

Usually It detected well though, when I was putting sentences from news outlets as a test.